Add an elevation prompt when trying to install system-wide packages
Add a generic mechanism for running specific code elevated. Due to the way the 'runas' verb works on Windows, this necessitates pipes for communication as the parent process cannot  communicate with the child with regular pipes when using 'UseShellExecute=true' which is required for verbs.

Adds the 'Run As Administrator' option to the 'Run Program' step. This is basically free for us to implement with the new abstraction because it works off of test steps.

Adds native FileLock mechanisms for Windows and Linux. This is required to ensure two different OpenTAP processes don't attempt to modify the system-wide directory at the same time.